[Update] Refactor main.js
- callback함수 리팩토링 시도중
