+++ shopping/public/index.html
@@ -5,7 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./stylesheets/style.scss" />
-    <script defer type="module" src="./javascripts/main.js"></script>
     <title>쇼핑하우 by kakaocommerce</title>
   </head>
 
@@ -857,6 +856,7 @@
     <a href="#top" class="btn_follow">
       <span><i class="fas fa-chevron-up"></i></span>
     </a>
+    <script type="module" src="./javascripts/main.js"></script>
     <script src="https://kit.fontawesome.com/439db4bb02.js" crossorigin="anonymous"></script>
   </body>
 </html>
